Moved science tests to separate case #22. Implemented doubles and quads too - closes #12
